Update README.md
# Data-Analysis--CO2-Emission-for-Canada-Vehicle
This project aims to predict the CO2 emissions of different vehicles based on their characteristics. 
The data used for this project is sourced from the first Edition of "Everyone Can Learn Python competition, 
which provides information on various features of different vehicles such as their engine size, fuel consumption, and transmission type.